App側LocationServiceの実装をファイル分離
